chore: moved methods to utils.rs
